Split src folder to render library and an example application. Not perfect, but better then it was.
