+++ CMakeLists.txt
@@ -6,28 +6,5 @@ project(
     LANGUAGES CXX
 )
 
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED True)
-
-include_directories(Eigen)
-include_directories(src)
-
-add_executable(
-    application
-    main.cpp
-    src/application.cpp
-    src/camera.cpp
-    src/hvertex.cpp
-    src/object.cpp
-    src/renderer.cpp
-    src/world.cpp
-    src/scenes.cpp
-    src/frame_movement.cpp
-)
-
-target_link_libraries(
-    application
-    sfml-graphics
-    sfml-window
-    sfml-system
-)
+add_subdirectory(Renderer3D)
+add_subdirectory(ExampleApplication)
